rename label file
